点类型改为double

+++ src/main/java/com/rast/Point2D.java
@@ -3,7 +3,7 @@ package com.rast;
 import com.demo.PixelDrawer;
 import com.demo.Point;
 
-public record Point2D(int x, int y, double h)
+public record Point2D(double x, double y, double h)
 {
 
     public static Point2D viewportToCanvas(double x ,double y)
@@ -12,7 +12,7 @@ public record Point2D(int x, int y, double h)
 
         double scaledY = y * RenderConfig.CANVAS_HEIGHT / RenderConfig.VIEW_SIZE;
 
-        return new Point2D((int) scaledX, (int) scaledY, 1.0);
+        return new Point2D( scaledX, scaledY, 1.0);
     }
 
     public static Point2D projectVertex(Point point)
